+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [3.5.5] 2020-04-23
+
 -   Report skipped tests correctly in junit xml
 -   Regard other tests as skipped when using only